# backend/apidoc/openapi.yaml
openapi: 3.0.0
info:
  description: API Documentation for Meal Planner App
  version: 1.0.0
  title: Meal Planner APIDoc
servers:
  - url: api
...
paths:
  /authors:
    post:
      tags:
        - authors
      summary: Add Author
      description: Adds an author
      operationId: addAuthor
      requestBody:
        description: An Author Object
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John
                link:
                  type: string
                  format: uri
                  example: "www.example.com/john"
              required:
                - name
      responses:
        "200":
          description: OK. Author succesfully added to database
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/authors/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "Author successfully added to database"
        "400":
          description: Bad Request. Incomplete Request Body
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Body should contain name"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /authors/{authorUID}:
    get:
      tags:
        - authors
      summary: Get Author By UID
      operationId: getAuthorByUID
      parameters:
        - in: path
          name: authorUID
          schema:
            type: string
          required: true
          description: The UID of the author
      responses:
        "200":
          description: OK. Author successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "John"
                      link:
                        type: string
                        format: uri
                        example: "www.example.com/john"
                  message:
                    type: string
                    example: "Author successfully retrieved"
        "404":
          description: Not Found. Author does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Author does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    put:
      tags:
        - authors
      summary: Update Author By UID
      description: Update Author Info By UID
      operationId: UpdateAuthorByUID
      parameters:
        - in: path
          name: userUID
          schema:
            type: string
          required: true
          description: The UID of the Author
      requestBody:
        description: An Author Object
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John
                link:
                  type: string
                  format: uri
                  example: "www.example.com/john"
              required:
                - name
      responses:
        "200":
          description: OK. Author succesfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "http://localhost:8080/api/authors/3PXW8J1QGkQlZgnjOaKv"
                  message:
                    type: string
                    example: "Author successfully updated!"
        "404":
          description: Not Found. Author does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Author does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
    delete:
      tags:
        - authors
      summary: Delete Author By UID
      description: Delete Author By UID
      operationId: deleteUserByUID
      parameters:
        - in: path
          name: authorUID
          schema:
            type: string
          required: true
          description: The UID of the Author
      responses:
        "200":
          description: OK. Author succesfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Author successfully deleted!"
        "404":
          description: Not Found. Author does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Author does not exist"
        "500":
          description: Internal Server Error. Some Error occured
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...